requirements.txt
requests>=2.31.0
anthropic>=0.39.0
openai>=1.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
